petverse/petverse-profiles/requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.35.1
gunicorn==20.1.0
google-cloud-aiplatform==1.54.0
shapely>=2.0.0